test: refactor define test. this is define package's tests
